requirements.txt
Flask==3.0.0
pymongo==4.6.1
cassandra-driver==3.29.0
python-dateutil==2.8.2
